ci: remove `shouldRunRelease` check
Previously, `shouldRunRelease` would check the commit type, and only run `release` if the type was either feat, fix, perf, or refactor. This check was only run on the latest commit, which works if the pr only includes a single commit or the top commit is within the range. There doesn't seem to be a way to get the commit range on circle, so this commit removes this check and will now run the release stage on every push to master