perf(angular): lazy load images
Previously, `ImageComponent` would always render the img in the DOM, even if it wasnt visible. This commit changes that to render the img in the DOM only if `LazyDirective` has indicated that it is visible.

This commit also adds a `load` event listener to the img, in order to react to when the browser loads the placeholder image and then the srcset images