refactor(angular): replace service error handlers with interceptor
Previously, each service had a `handleError` function that logged the error and returned a default safe value to the subscriber. This meant that every http function would have to use `catchError` to call the `handleError` function, and it also meant that errors were being swallowed before they got to the subscriber. This commit removes the `handleError` function and replaces it with a http interceptor, which catches and rethrows the error to the subscriber