refactor(angular): replace custom transfer state implementation
Previously, functions in http services would individually get and set the `TransferState` cache. This logic meant that the cache would not only be hit going from server -> browser, but it would also be hit for client side routing (e.g. work -> about -> work).

This commit removes that implementation and replaces it with the generic `TransferHttpCacheModule` for the sake of reducing complexity, which installs an interceptor and handles the cache itself. This does mean that there will no longer be any cache hits when angular has bootstrapped client-side, however.