refactor(angular): remove resolver services
Previously, each routed component had a resolver service that would fetch data the component needed before the navigation happened. This meant that if a route or data did not exist, the resolver service would not route to the component, and would handle the situation itself. For users with slow network connections, it would appear as though nothing was happening.

This commit removes the resolver services and instead makes the component fetch the data it needs itself. This means that routing happens instantly, and this should make it so each component can display its own loading or error notifications in the future