feat(angular): handle network errors
Previously, an error in a network subscription would bubble up to the global error handler as it was not caught, which would display the generic `ErrorComponent`. This commit changes this by handling the error in the component using a custom `catchNetworkError` pipe. Instead of the previous generic error message, the user is offered the option to retry the request

NOTE: This solution was duplicated in each component and not implemented as an interceptor because of requests such as the form submission which needs to display a custom message