fix(angular): handle network errors loading more news posts
Previously, angular used a load more button to display more news posts. This button incremented the current page number saved locally in `PrismicService`, and on news -> news-post, `PrismicService` would send one network request to fetch all the posts you had previously loaded. So if you were on page 2 and the number of posts displayed were 9, it would fetch a page 1 of size 18. This meant that there was only 1 network request to get back to page 2 instead of 2 separate calls of size 9, and it meant that the user didnt lose progression. This method worked fine for when every network request was successful, but for situations where a load more request failed, the page size and page number state would be wrong.
This commit removes all this logic and simplifies the implentation by adding page numbers to request urls and letting the router manage state.
Note: `RouteReuseStrategy` was considered but decided against, as it would be adding extra complexity