refactor(angular): use `async` pipe
Previously, `PrismicService` `getPost` returned either `NewsPost`, `CareerPost`, or `null`. That meant that subscribers to `getPost` could not use `async` and instead had to subscribe manually, as `ngIf` would always return false for `null`. So, it was impossible to tell whether `post$` was just `undefined` or whether `getPost` actually returned `null`, and so we should show a 404 banner.
This commit changes `getPost` to return results wrapped in an object instead so that `async` can be used, and removes the manual subscription in `CareerComponent` and `NewsComponent`