refactor: only use initial route snapshot
Previously `CareerComponent`, `NewsPostComponent`, and `PostComponent` subscribed to the route `paramMap`, using a `switchMap` to fetch the latest post when route params changed. This meant that if a user routed to the same component, angular would re-use the component instead of destroying it and creating another. However, currently there is no way to route to the same component, and so this commit replaces this and instead gets the initial params only